refactor trust changeset name
