+++ src/rpm.rs
@@ -1,6 +1,6 @@
-use crate::api;
-use crate::fapolicyd::keep_entry;
-use crate::rpm::RpmError::{Discovery, Execution, NotFound};
+use std::process::Command;
+
+use nom::{InputIter, Parser};
 use nom::bytes::complete::tag;
 use nom::character::complete::alphanumeric1;
 use nom::character::complete::digit1;
@@ -8,8 +8,10 @@ use nom::character::complete::line_ending;
 use nom::character::complete::space1;
 use nom::combinator::iterator;
 use nom::sequence::{delimited, terminated};
-use nom::{InputIter, Parser};
-use std::process::Command;
+
+use crate::api;
+use crate::fapolicyd::keep_entry;
+use crate::rpm::RpmError::{Discovery, Execution, NotFound};
 
 #[derive(Debug)]
 struct RpmDbEntry {
